create ContentLayout and reuse it in upgrade, roadmap, staking, translatathon layouts
