+++ src/components/Hero/CallToAction.tsx
@@ -1,6 +1,10 @@
 import type { ReactNode } from "react"
 
-import { Button, type ButtonProps } from "@/components/ui/buttons/Button"
+import {
+  Button,
+  ButtonLink,
+  type ButtonProps,
+} from "@/components/ui/buttons/Button"
 
 import { cn } from "@/lib/utils/cn"
 import { type MatomoEventOptions, trackCustomEvent } from "@/lib/utils/matomo"
@@ -10,8 +14,9 @@ export type CallToActionProps = Omit<
   "children" | "content" | "variant" | "isSecondary"
 > & {
   content: ReactNode
-  matomo: MatomoEventOptions
+  matomo?: MatomoEventOptions
   index: number
+  href?: string
 }
 
 export const CallToAction = ({
@@ -19,14 +24,26 @@ export const CallToAction = ({
   matomo,
   index,
   className,
+  href,
   ...props
 }: CallToActionProps) => {
-  const handleClick = () => trackCustomEvent(matomo)
+  const handleClick = () => matomo && trackCustomEvent(matomo)
 
   const buttonProps: ButtonProps = {
     variant: index === 0 ? "solid" : "outline",
     isSecondary: index !== 0,
     className: cn("flex-[1] md:flex-[initial]", className),
+  }
+
+  if (href) {
+    return (
+      <ButtonLink
+        href={href}
+        buttonProps={{ onClick: handleClick, ...buttonProps, ...props }}
+      >
+        {content}
+      </ButtonLink>
+    )
   }
 
   return (
