+++ src/components/LeftNavBar/index.tsx
@@ -1,68 +1,47 @@
-import { calc, Flex, type FlexProps, type HeadingProps } from "@chakra-ui/react"
-
 import type { ToCItem } from "@/lib/types"
 
 import ButtonDropdown, {
-  ButtonDropdownProps,
   List as ButtonDropdownList,
 } from "@/components/ButtonDropdown"
-import OldHeading from "@/components/OldHeading"
 import Translation from "@/components/Translation"
 import UpgradeTableOfContents from "@/components/UpgradeTableOfContents"
 
-export const H2 = (props: HeadingProps) => (
-  <OldHeading
-    fontWeight={700}
-    lineHeight={1.4}
-    size="lg"
-    textAlign="start"
-    mt={0}
-    {...props}
-  />
-)
+import { cn } from "@/lib/utils/cn"
 
-export const StyledButtonDropdown = ({
-  list,
-  ...rest
-}: FlexProps & Pick<ButtonDropdownProps, "list">) => (
-  <Flex align="flex-end" justify="flex-end" mb={8} {...rest} pos="relative">
-    <ButtonDropdown list={list} w="full" minW="240px" />
-  </Flex>
-)
-
-type LeftNavBarProps = FlexProps & {
+export type LeftNavBarProps = {
   dropdownLinks?: ButtonDropdownList
   maxDepth?: number
   tocItems: ToCItem[]
+  className?: string
 }
 
 const LeftNavBar = ({
   dropdownLinks,
   maxDepth = 1,
   tocItems,
+  className,
   ...props
 }: LeftNavBarProps) => {
   return (
-    <Flex
-      as="aside"
-      flexDirection="column"
-      flex="0 1 400px"
-      ms={8}
-      me={16}
-      position="sticky"
-      top="6.25rem"
-      height={calc("100vh").subtract("80px").toString()}
-      zIndex={99}
+    <aside
+      className={cn(
+        "z-99 sticky top-[6.25rem] me-16 ms-8 flex h-[calc(100vh-80px)] basis-[400px] flex-col",
+        className
+      )}
       {...props}
     >
-      {dropdownLinks && <StyledButtonDropdown list={dropdownLinks} />}
-      <H2>
+      {dropdownLinks && (
+        <div className="relative mb-8 flex items-end justify-end">
+          <ButtonDropdown list={dropdownLinks} w="full" minW="240px" />
+        </div>
+      )}
+      <h2 className="mb-8 text-3xl leading-xs">
         <Translation id="on-this-page" />
-      </H2>
+      </h2>
       {tocItems && (
         <UpgradeTableOfContents items={tocItems} maxDepth={maxDepth} />
       )}
-    </Flex>
+    </aside>
   )
 }
