+++ src/layouts/index.ts
@@ -1,9 +1,9 @@
 export * from "./BaseLayout"
 export * from "./Docs"
-export * from "./Roadmap"
-export * from "./Staking"
+export * from "./md"
+export * from "./md/Roadmap"
+export * from "./md/Staking"
+export * from "./md/Translatathon"
 export * from "./Static"
-export * from "./Translatathon"
 export * from "./Tutorial"
-export * from "./Upgrade"
 export * from "./UseCases"
