+++ src/layouts/md/Upgrade.tsx
@@ -1,53 +1,18 @@
 import { useTranslation } from "next-i18next"
-import {
-  Box,
-  type BoxProps,
-  type FlexProps,
-  List,
-  ListItem,
-  Text,
-  useToken,
-} from "@chakra-ui/react"
 
 import type { ChildOnlyProp } from "@/lib/types"
 import type { MdPageContent, UpgradeFrontmatter } from "@/lib/interfaces"
 
 import BeaconChainActions from "@/components/BeaconChainActions"
 import type { List as ButtonDropdownList } from "@/components/ButtonDropdown"
-import FeedbackCard from "@/components/FeedbackCard"
-import { ContentHero } from "@/components/Hero"
-import LeftNavBar from "@/components/LeftNavBar"
-import {
-  ContentContainer,
-  MobileButton,
-  MobileButtonDropdown,
-  Page as MdPage,
-} from "@/components/MdComponents"
 import MergeArticleList from "@/components/MergeArticleList"
 import MergeInfographic from "@/components/MergeInfographic"
+import { List, ListItem } from "@/components/ui/list"
 import UpgradeStatus from "@/components/UpgradeStatus"
 
 import { getSummaryPoints } from "@/lib/utils/getSummaryPoints"
 
-const Page = (props: FlexProps) => <MdPage sx={{}} {...props} />
-
-type ContainerProps = Pick<BoxProps, "children" | "dir">
-
-const Container = (props: ContainerProps) => (
-  <Box position="relative" {...props} />
-)
-
-const LastUpdated = (props: ChildOnlyProp) => (
-  <Text
-    color="text200"
-    fontStyle="italic"
-    pt={4}
-    mb={0}
-    borderTop="1px"
-    borderColor="body.dark"
-    {...props}
-  />
-)
+import { ContentLayout } from "../ContentLayout"
 
 // Upgrade layout components
 export const upgradeComponents = {
@@ -101,47 +66,33 @@ export const UpgradeLayout = ({
     ],
   }
 
-  const lgBreakpoint = useToken("breakpoints", "lg")
-
   return (
-    <Container dir={contentNotTranslated ? "ltr" : "unset"}>
-      <ContentHero
-        breadcrumbs={{ slug, startDepth: 1 }}
-        title={frontmatter.title}
-        description={
+    <ContentLayout
+      dir={contentNotTranslated ? "ltr" : "unset"}
+      heroProps={{
+        ...frontmatter,
+        breadcrumbs: { slug, startDepth: 1 },
+        heroImg: frontmatter.image,
+        description: (
           <>
-            <Box>
-              <List listStyleType="disc">
+            <div>
+              <List>
                 {summaryPoints.map((point, idx) => (
                   <ListItem key={idx}>{point}</ListItem>
                 ))}
               </List>
-            </Box>
+            </div>
 
-            <LastUpdated>
+            <p className="border-t pt-4 italic">
               {t("common:page-last-updated")}: {lastEditLocaleTimestamp}
-            </LastUpdated>
+            </p>
           </>
-        }
-        heroImg={frontmatter.image}
-        blurDataURL={frontmatter.blurDataURL}
-      />
-      <Page>
-        {/* TODO: Switch to `above="lg"` after completion of Chakra Migration */}
-        <LeftNavBar
-          hideBelow={lgBreakpoint}
-          dropdownLinks={dropdownLinks}
-          tocItems={tocItems}
-          maxDepth={frontmatter.sidebarDepth!}
-        />
-        <ContentContainer>
-          {children}
-          <FeedbackCard />
-        </ContentContainer>
-        <MobileButton>
-          <MobileButtonDropdown list={dropdownLinks} />
-        </MobileButton>
-      </Page>
-    </Container>
+        ),
+      }}
+      tocItems={tocItems}
+      dropdownLinks={dropdownLinks}
+    >
+      {children}
+    </ContentLayout>
   )
 }
